requirements.txt
asgiref==3.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==6.0.0
Django==3.2.3
django-environ==0.10.0
django-restframework==0.0.1
djangorestframework==3.14.0
environ==1.0
Faker==23.1.0
idna==3.6
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.6
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.2.0
